Fixes two merge bugs:
1. The whole_state struct now keeps an array of window sizes so that
during reconstruction it can use the same window size as the target
delta.  The code was previously using dec_tgtlen, which was the window
size of the last target delta window.

2. xd3_merge_copy_source(), which applys a source-copy instruction
during merge was not properly translating target-copy instructions
in the target.  The solution here is SLOW and INEFFICIENT, but it at
least allows the tests to pass.  A big TODO here is to improve the
algorithm: it has a potentially O(N) recursion for each target-copy
that it sees, and the naive approach also can produce duplicate adds.